separate logic from DOM
