+++ src/js/content.js
@@ -1,3 +1,6 @@
+// import switchScale from './apii';
+// import swiitchScale from './apii';
+
 const userInput = () => {
   const formContainer = document.createElement('div');
   formContainer.classList.add('flex', 'justify-around', 'pt-8');
@@ -33,4 +36,110 @@ const userInput = () => {
   return formContainer;
 };
 
-export default userInput;
+const weatherCard = () => {
+  const root = document.querySelector('#root');
+  // const { main } = data
+  // card.innerHTML = '';
+  // let show;
+
+  // switchScale(data);
+  
+
+  const card = document.createElement('div');
+  card.classList.add('pt-6', 'flex', 'flex-row', 'content-center');
+
+  const cardContainer = document.createElement('div');
+  cardContainer.classList.add('p-4', 'w-5/6', 'mx-auto', 'px-auto', 'text-center');
+
+  // City time
+  const cardTime = document.createElement('h2');
+  cardTime.classList.add('text-2xl', 'text-white', 'card-time');
+  // cardTime.innerHTML = `<h2>${getLocalTime(data)}</h2>`;
+
+  // City Name
+  const cardCity = document.createElement('h2');
+  cardCity.classList.add('text-4xl', 'text-blue-900', 'card-city');
+  // cardCity.innerText = `${data.name}, ${data.sys.country}`;
+
+  // City properties container
+  const cardInfo = document.createElement('div');
+  cardInfo.classList.add('flex', 'justify-around', 'pt-6');
+
+  const cardTempDiv = document.createElement('div'); // Temperature
+
+  const cardTemp = document.createElement('h2');
+  cardTemp.classList.add('text-2xl', 'text-white', 'font-thin');
+  cardTemp.innerHTML = '<h2>Temperature</h2>';
+
+  const temp = document.createElement('h2');
+  temp.classList.add('text-3xl', 'card-temp');
+  // temp.innerHTML = `<h2>${Math.round(data.main.temp - 273.15)}  &#176;C</h2>`;
+
+  const cardHumDiv = document.createElement('div'); // Humidity
+
+  const cardHum = document.createElement('h2');
+  cardHum.classList.add('text-2xl', 'text-white');
+  cardHum.innerHTML = '<h2>Humidity</h2>';
+
+  const humi = document.createElement('h2');
+  humi.classList.add('text-3xl', 'humidity');
+  // humi.innerHTML = `<h2>${data.main.humidity}</h2>`;
+
+  const cardPreDiv = document.createElement('div'); // Pressure
+
+  const cardPressure = document.createElement('h2');
+  cardPressure.classList.add('text-2xl', 'text-white');
+  cardPressure.innerHTML = '<h2>Pressure</h2>';
+
+  const pressure = document.createElement('h2');
+  pressure.classList.add('text-3xl', 'pressure');
+  // pressure.innerHTML = `<h2>${data.main.pressure}</h2>`;
+
+  const cardIconDiv = document.createElement('div'); // Wind speed
+
+  const cardIcon = document.createElement('h2');
+  cardIcon.classList.add('text-2xl', 'text-white');
+  cardIcon.innerHTML = 'Wind Speed';
+
+  const speed = document.createElement('h2');
+  speed.classList.add('text-3xl', 'speed');
+  // speed.innerHTML = `${data.wind.speed} M/H`;
+
+  const cardCloudDiv = document.createElement('div'); // Description
+
+  const cardCloud = document.createElement('h2');
+  cardCloud.classList.add('text-2xl', 'text-white');
+  cardCloud.innerText = 'Cloud description';
+
+  const cloud = document.createElement('h2');
+  cloud.classList.add('text-3xl', 'description');
+  // cloud.innerText = `${data.weather[0].description}`;
+
+  cardTempDiv.appendChild(cardTemp);
+  cardTempDiv.appendChild(temp);
+
+  cardHumDiv.appendChild(cardHum);
+  cardHumDiv.appendChild(humi);
+
+  cardPreDiv.appendChild(cardPressure);
+  cardPreDiv.appendChild(pressure);
+
+  cardIconDiv.appendChild(cardIcon);
+  cardIconDiv.appendChild(speed);
+
+  cardCloudDiv.appendChild(cardCloud);
+  cardCloudDiv.appendChild(cloud);
+
+  cardContainer.appendChild(cardTime);
+  cardContainer.appendChild(cardCity);
+  cardInfo.appendChild(cardTempDiv);
+  cardInfo.appendChild(cardHumDiv);
+  cardInfo.appendChild(cardPreDiv);
+  cardInfo.appendChild(cardIconDiv);
+  cardInfo.appendChild(cardCloudDiv);
+  cardContainer.appendChild(cardInfo);
+  card.appendChild(cardContainer);
+  root.appendChild(card);
+};
+
+export { userInput, weatherCard };
